app/requirements.txt
fastapi
uvicorn
psycopg[binary]
requests
python-dotenv